Add changes for frame forwarding and inversion
While this version should work as intended. There is an implicit override in firmware where if you attempt to forward frame0 and frame1 to the same tone, frame0 takes precedence. This should raise a warning such that users are not caught off guard, or we should potentially change the format for the input arguments such that there is no room for confusion. Just keep in mind that conflicting data won't cause issues on the hardware for the current state fo the code.