Soft LUT-based streaming
When falling back on streaming (for the whole sequence) in the event of
a LUT overflow, use the intermediate LUTs in software to regenerate the
sequence as opposed to calculating the streaming representation directly
from gate slices. This gives nearly equivalent times (maybe 5% longer in
the LUT-based scheme) for generating streaming bytecode up front and
attempting compression and falling back on streaming with the soft LUTs.