Add subcircuit handling for interleaved streaming
This version works against HW with initial r/w counter implementation
for stabilizing busy signals through the sequence pipeline (despite the
fact that the test build DNMT*). Reduced the number of index operations
on tuples/lists by replacing lmdatoset with a defaultdict(list) and
nplutls with a dict. These were eating into 20% of compile time for a
test file with ~1e4 unique gates, 2/3 of which use 26 knots for spline
modulation. The compile time (including gatelet distillation and
interleaved streaming for packet generation) for this file is ~5
seconds, and is now 40% parsing/visiting, and 40% generate_lookup_tables
(roughly 30% binarization and 10% recasting, where 5% is spline fitting
and mapping when using a custom Go .so).

*However in this case latency incurred by the signals with WNS should
not impact performance in principle, unless routing optimization
sacrificed timing elsewhere. But everything appears to be working as
expected between SW emulation, FW simulation, and HW.

Clean out some dead code and print statements

Clean up and consolidate code