Add optional phase inversion parameter
This is to avoid carrying around local changes on a per-setup basis. The
basic issue has to do with the "natural" sign for phase, which may be
positive or negative depending on whether phase is applied to the
upper/lower frequency tone for a Raman pulse.

USAGE At the beginning of a pulse definition file add the following two
lines:

    import jaqalpaw.utilities.parameters
    jaqalpaw.utilities.parameters.INVERT_PHASE = True

otherwise, INVERT_PHASE defaults to False.